claire/internal-asset-search
